// src/__generated__/merge-pull-request.ts
/* eslint-disable */
/* This is an autogenerated file. Do not edit this file directly! */
type MergePullRequestInputInputType = {
    authorEmail?: string | null;
    clientMutationId?: string | null;
    commitBody?: string | null;
    commitHeadline?: string | null;
    expectedHeadOid?: any | null;
    mergeMethod?: ("MERGE" | "REBASE" | "SQUASH") | null;
    pullRequestId: string;
};
export type MergePullRequest = {
    mergePullRequest: ({
        pullRequest: ({
            merged: boolean;
            mergedAt: any | null;
            state: "CLOSED" | "MERGED" | "OPEN";
            url: any;
        }) | null;
    }) | null;
};
export type MergePullRequestVariables = {
    input: MergePullRequestInputInputType;
};