// src/__generated__/update-issue-input.ts
/* eslint-disable */
/* This is an autogenerated file. Do not edit this file directly! */
export type UpdateIssueInput = {
    updateIssue: ({
        issue: ({
            id: string;
            closedAt: any | null;
            state: "CLOSED" | "OPEN";
        }) | null;
    }) | null;
};
export type UpdateIssueInputVariables = {
    id: string;
    state: "CLOSED" | "OPEN";
};